pyproject.toml
[project]
name = "dais-sdk"
authors = [{ name = "BHznJNs", email = "bhznjns@outlook.com" }]
description = "A LLM Agent Meta Framework"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.14",
]
requires-python = ">=3.14"
version = "0.7.5"

dependencies = [
    "openai>=2.24.0",
    "pydantic>=2.0.0",

    # MCP client dependencies
    "httpx>=0.28.0",
    "mcp>=1.26.0",
    "starlette>=0.50.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = ["python-dotenv>=1.2.1"]
test = [
    "pytest-cov",
    "pytest-mock",
    "pytest-runner",
    "pytest-asyncio",
    "pytest",
    "pytest-github-actions-annotate-failures",
]

[project.urls]
Source = "https://github.com/Dais-Project/Dais-SDK"
Tracker = "https://github.com/Dais-Project/Dais-SDK/issues"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"

[tool.pyright]
include = ["src"]
pythonVersion = "3.14"
typeCheckingMode = "standard"

[tool.pytest.ini_options]
addopts = "--cov-report xml:coverage.xml --cov src --cov-fail-under 0 --cov-append -m 'not integration'"
pythonpath = ["src"]
testpaths = "tests"
markers = [
    "integration: marks as integration test",
    "notebooks: marks as notebook test",
    "gpu: marks as gpu test",
    "spark: marks tests which need Spark",
    "slow: marks tests as slow",
    "unit: fast offline tests",
]
